build: resolve 1 instance of -Wsign-compare
lib/contexts_pool.cpp:202:30: warning: comparison between signed and unsigned
integer expressions [-Wsign-compare]
   for (unsigned int i = 0; i < num; ++i) {

+++ lib/contexts_pool.cpp
@@ -199,7 +199,7 @@ static void *ctxp_thrwork(void *pparam)
 		if (num <= 0) {
 			continue;
 		}
-		for (unsigned int i = 0; i < num; ++i) {
+		for (unsigned int i = 0; i < static_cast<unsigned int>(num); ++i) {
 			auto pcontext = g_poll_ctx.get_data(i);
 			std::unique_lock poll_hold(g_context_locks[CONTEXT_POLLING]);
 			if (CONTEXT_POLLING != pcontext->type) {
